Implement turn-based system in rugby and football games
Added a variable 'is_rugby_turn' to control which game gets the first turn and toggles between the two games at the end of each one. Also, updated the conditions for starting each game and generated random number of players to start the rugby game. Thereby, creating a more dynamic and unpredicted pattern of the games.